# Update: Less pointers & indirections