--- a/makefile
+++ b/makefile
@@ -6,6 +6,7 @@ rfc5424/machine.go: rfc5424/machine.go.rl
 .PHONY: build
 build: rfc5424/machine.go
 	@gofmt -w -s $<
+	@sed -i '/^\/\/line/d' $<
 
 .PHONY: bench
 bench: rfc5424/*_test.go rfc5424/machine.go