… rfc5424/machine.go