The code within this PR does 2 things:

1. Tells ragel to do not generate status variables we do not use
2. Brings out the output message from the internal machine

Point 2 slightly affects the performances.

```
[ok]_with_escaped_backslash_within_structured_data-4     1592          1442          -9.42%
[ok]_with_multiple_structured_data________________-4     1665          1585          -4.80%
[ok]_very_long_message____________________________-4     4432          4252          -4.06%
[ok]_minimal_with_message_containing_newline______-4     366           379           +3.55%
[ok]_with_structured_data_id,_w/o_structured_data_-4     1072          1039          -3.08%
[no]_impossible_timestamp_________________________-4     1150          1123          -2.35%
[no]_malformed_structured_data____________________-4     586           573           -2.22%
[no]_empty_input__________________________________-4     234           239           +2.14%
[ok]_average_message______________________________-4     2231          2187          -1.97%
[ok]_minimal_with_UTF-8_message___________________-4     566           558           -1.41%
[ok]_minimal______________________________________-4     301           305           +1.33%
[ok]_all_max_length_and_complete__________________-4     3071          3111          +1.30%
[no]_multiple_syslog_messages_on_multiple_lines___-4     484           478           -1.24%
[ok]_w/o_procid,_w/o_structured_data,_with_message-4     1037          1027          -0.96%
[no]_with_duplicated_structured_data_id___________-4     1481          1467          -0.95%
[ok]_with_UTF-8_structured_data_param_value,_with_-4     1574          1588          +0.89%
[ok]_all_max_length_except_structured_data_and_mes-4     1985          1994          +0.45%
[ok]_complicated_message__________________________-4     1840          1834          -0.33%
```

I'd like to hear from you guys @goller and @fntlnz.