@leodido 2018-04-27
Latest commits achieve another huge impact on performances (both ns/op & memory allocations).

Those commits introduce and use an intermediate pvt struct within which the machine stores data.

The hydration within the final `SyslogMessage` is done at the end with the help of some smart flags within the intermediate instance.

The API is unchanged and the semantic meaning of nil values for public `SyslogMessage` is preserved.

So this PR does also:

3. Reducing/removing pointers not changing the meaning of the `SyslogMessage`'s field neither the public API

Following the performance gains (related to previous one, not to the develop branch).

```
                                                         old ns/op   new ns/op   delta
[no]_empty_input__________________________________-4     280         253         -9.64%
[no]_multiple_syslog_messages_on_multiple_lines___-4     590         433         -26.61%
[no]_impossible_timestamp_________________________-4     1408        1080        -23.30%
[no]_malformed_structured_data____________________-4     651         552         -15.21%
[no]_with_duplicated_structured_data_id___________-4     1643        1246        -24.16%
[ok]_minimal______________________________________-4     504         264         -47.62%
[ok]_average_message______________________________-4     2628        1984        -24.51%
[ok]_complicated_message__________________________-4     1953        1644        -15.82%
[ok]_very_long_message____________________________-4     4838        3826        -20.92%
[ok]_all_max_length_and_complete__________________-4     4548        2792        -38.61%
[ok]_all_max_length_except_structured_data_and_mes-4     2900        1830        -36.90%
[ok]_minimal_with_message_containing_newline______-4     442         294         -33.48%
[ok]_w/o_procid,_w/o_structured_data,_with_message-4     1134        956         -15.70%
[ok]_minimal_with_UTF-8_message___________________-4     665         586         -11.88%
[ok]_with_structured_data_id,_w/o_structured_data_-4     1598        998         -37.55%
[ok]_with_multiple_structured_data________________-4     2085        1538        -26.24%
[ok]_with_escaped_backslash_within_structured_data-4     1722        1316        -23.58%
[ok]_with_UTF-8_structured_data_param_value,_with_-4     2758        1580        -42.71%

                                                         old allocs  new allocs  delta
[no]_empty_input__________________________________-4     3           3           +0.00%
[no]_multiple_syslog_messages_on_multiple_lines___-4     15          12          -20.00%
[no]_impossible_timestamp_________________________-4     17          11          -35.29%
[no]_malformed_structured_data____________________-4     15          12          -20.00%
[no]_with_duplicated_structured_data_id___________-4     29          17          -41.38%
[ok]_minimal______________________________________-4     12          9           -25.00%
[ok]_average_message______________________________-4     35          26          -25.71%
[ok]_complicated_message__________________________-4     34          25          -26.47%
[ok]_very_long_message____________________________-4     40          28          -30.00%
[ok]_all_max_length_and_complete__________________-4     40          28          -30.00%
[ok]_all_max_length_except_structured_data_and_mes-4     23          13          -43.48%
[ok]_minimal_with_message_containing_newline______-4     14          10          -28.57%
[ok]_w/o_procid,_w/o_structured_data,_with_message-4     19          11          -42.11%
[ok]_minimal_with_UTF-8_message___________________-4     14          10          -28.57%
[ok]_with_structured_data_id,_w/o_structured_data_-4     22          14          -36.36%
[ok]_with_multiple_structured_data________________-4     30          22          -26.67%
[ok]_with_escaped_backslash_within_structured_data-4     29          20          -31.03%
[ok]_with_UTF-8_structured_data_param_value,_with_-4     30          21          -30.00%
```

One image better than a thosand words.

![benchviz](https://user-images.githubusercontent.com/120051/39341709-b7df173a-49d4-11e8-8322-0f0815334c81.png)